requirements.txt
absl-py==2.2.0
astunparse==1.6.3
async-timeout==4.0.3
audioread==3.0.1
beautifulsoup4==4.13.3
bleak==0.22.3
bleak-winrt==1.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
facenet-pytorch==2.6.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
FoxDot==0.8.12
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==5.29.4
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-osc==1.8.3
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
supercollider==0.0.6
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webrtcvad==2.0.10
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
fastDTW>=0.3.4
